requirements.txt
amqp==2.4.2
attrdict==2.0.1
Babel==2.6.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
cytoolz==0.9.0.1
Django==2.2.1
djangorestframework==3.9.3
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.2
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.5.2
flower==0.9.3
gunicorn==19.9.0
hexbytes==0.1.0
idna==2.8
kombu==4.5.0
lru-dict==1.1.6
parsimonious==0.8.1
psycopg2==2.8.2
pycryptodome==3.8.1
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
rlp==1.1.0
six==1.12.0
sqlparse==0.3.0
toolz==0.9.0
tornado==5.1.1
urllib3==1.24.3
vine==1.3.0
web3==4.9.2
websockets==6.0